+++ chrome/css/setting.css
@@ -1,0 +1,110 @@
+#setting_div{
+    border: 3px solid #999;
+    background: #FFF;
+    width: 568px;
+    box-shadow: 0 0 9px #999;
+}
+#setting_div_table input{
+    text-indent: 10px;
+    border: 1px solid #C6C6C6;
+    box-shadow: 0 0 3px #C6C6C6;
+    -webkit-box-shadow: 0 0 3px #C6C6C6;
+}
+.input-large{
+    width:85%;
+}
+.input-medium{
+    width:81%;
+}
+
+.input-small{
+    width:150px;
+}
+#setting_div_table input[disabled]{
+    cursor: not-allowed;
+    background-color: #eee;
+}
+#send_test{
+    display:inline-block;
+    border:1px solid #D1D1D1;
+    background-color: #F7F7F7;
+    text-align: center; text-decoration: none;
+    color:#1B83EB;
+}
+#add_rpc{
+    display:inline-block;
+    border:1px solid #D1D1D1;
+    background-color: #F7F7F7;
+    text-align: center; text-decoration: none;
+    color:#1B83EB;
+    margin-left:5px;
+}
+#copyright{
+    display:inline-block;
+}
+#setting_aria2_headers{
+    overflow:auto;
+    resize:none;
+    width:90%;
+    height:80px;
+    border: 1px solid #C6C6C6;
+    box-shadow: 0 0 3px #C6C6C6;
+    -webkit-box-shadow: 0 0 3px #C6C6C6;
+}
+a{
+    color: #666 !important;
+    text-decoration: none !important;
+}
+.button{
+    background: #eeeeef;
+    margin: 0 10px;
+    text-decoration: none;
+    font-size: 14px;
+    padding:0 50px;
+    height: 37px;
+    line-height: 37px;
+    color: #333;
+    border: 1px solid rgb(198, 198, 198);
+    border-radius: 2px;
+    display: inline-block;
+}
+.button:hover,.button:visited,.button:active{
+    background: #fafafa;
+    color: #333;
+    text-decoration: none;
+}
+.new-dbtn .menu{
+    position: absolute;
+    width: 98%;
+    left: 0;
+    top: 32px;
+    background: #fff;
+    text-align: center;
+    border: 1px solid #aaa;
+    display: none;
+    z-index:100;
+}
+.new-dbtn .menu a{
+    height: 24px;
+    line-height: 24px;
+    display: block;
+    color: #666;
+}
+.new-dbtn .menu a:hover{
+    background: #e4eefe;
+    cursor: pointer;
+    text-decoration: none;
+}
+.download-mgr-dialog .content {
+    padding: 10px 20px;
+    height:auto;
+    overflow: hidden;
+}
+#setting_div_table{
+    width:100%;
+    border:0;
+    border-collapse:separate;
+    border-spacing:10px;
+    display:table;
+    background-color: rgb(250, 250, 250);
+}
